requirements.txt
requests
azure-storage-blob==12.3.2
lxml
sqlalchemy
psycopg2-binary